+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+ - Include generated Github release notes in the release description
+ - Install instructions in the docs
+
+### Fixed 
+
  - Fixed update_pod_label subcommand functionality for service.
 
 ## [0.2.33] - 2024-08-12